Update: Move project files into `kodi_strm` package
Move existing code files into a package named `kodi_strm` — this
allows using the project by allowing a call to the package be a
replacment for making a call to `strm-generator.py`
