+++ kodi_strm/__main__.py
@@ -1,4 +1,4 @@
-import kodi_strm.cli as cli
+from . import cli
 
 if __name__ == "__main__":
     cli.main()
